echo_doc/requirements.txt
streamlit
pinecone
langchain-community
langchain-text-splitters
python-dotenv
cohere
pdfplumber